+++ HISTORY.md
@@ -51,6 +51,7 @@
 * objects, extensions: added support for pen modes ("paint", "erase" and "create")
 * objects: support stamping on sprites using pen modes
 * objects: refactored pen blendingMode
+* objects: refactored pen surface architecture
 
 ### 2025-12-10
 * gui: tweaked IDE resizing layout for different translations